[Backport 1.9] COMPASS 1763 Geo query circle does not show at first after query is executed (#1218)
* :art: Rename usage of <UniqueMinichart> to <UniqueMiniChart>

* :art: Make QueryStore listener reusable

* :bug: Populate query initially

* :art: Rename query to queryValue within the <UniqueMinichart> hierarchy

Note: I'm skipping over the <D3Component> hierarchy as it's not currently affected.

* :art: Extract <ValueBubble> into its own component file

* :hammer: Rename unique.jsx to unique-minichart.jsx
